requirements.txt
pandas
numpy
matplotlib
seaborn
python-pptx

!python-pptx